+++ src/app/pages/cases/create/CreateCasePage.tsx
@@ -11,7 +11,7 @@ import AddressHeadingByBagId from "app/components/shared/AddressHeadingByBagId/A
 import CreateForm from "app/components/cases/CreateForm/CreateForm"
 import NotFoundPage from "app/pages/errors/NotFoundPage"
 import { Column } from "app/components/layouts/Grid"
-import NoValidTonIdAlert from "./NoValidTonIdAlert"
+import InvalidTonIdAlert from "./InvalidTonIdAlert"
 
 type Props = {
   bagId: string
@@ -21,12 +21,12 @@ type Props = {
 const CreateCasePage: React.FC<RouteComponentProps<Props>> = ({ bagId }) => {
   const tonIdParam = getUrlParam("tonId")
   const tonId = parseUrlParamId(tonIdParam)
-  const isNoValidTonId = tonIdParam !== undefined && tonId === undefined
+  const isInvalidTonId = tonIdParam !== undefined && tonId === undefined
 
   return (
   isValidUrlParamBAGId(bagId) ?
     <DefaultLayout>
-      <NoValidTonIdAlert isVisible={ isNoValidTonId } />
+      { isInvalidTonId && <InvalidTonIdAlert /> }
       <RowWithColumn>
         <PageHeading />
       </RowWithColumn>
